feat: 支持github公开仓库镜像拉取 #65
